# lib/api-zod/src/generated/types/createResourceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Geo3D Library API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateResourceInputType } from "./createResourceInputType";

export interface CreateResourceInput {
  title: string;
  description?: string;
  type: CreateResourceInputType;
  categoryId?: number;
  region?: string;
  author?: string;
  year?: number;
  thumbnailUrl?: string;
  downloadUrl?: string;
  tags?: string[];
}