# lib/api-zod/src/generated/types/resourceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Geo3D Library API
 * OpenAPI spec version: 0.1.0
 */
import type { Resource } from "./resource";

export interface ResourceListResponse {
  resources: Resource[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}